requirements.txt
# OpenAI ve LangChain
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10

# data fetching
yfinance>=0.2.18
pycoingecko>=3.1.0

# Vector store ve embeddings
faiss-cpu>=1.7.4

# Environment variables
python-dotenv>=1.0.0

# Flask Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# other helper libraries
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
